moving component files to another folder
