+++ src/items/components/Descriptor.ts
