+++ src/items/components/Inventory.ts
