+++ src/items/components/PickedUp.ts
