+++ src/items/components/Valuable.ts
